route helper function moved out of menu into global helpers
